LineLayout sample ported to 64-bits

+++ LineLayout/LineLayout/LineLayoutViewController.cs
@@ -1,7 +1,7 @@
 using System;
 
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 
 namespace LineLayout
 {
@@ -18,7 +18,7 @@ namespace LineLayout
 			CollectionView.RegisterClassForCell (typeof(Cell), cellToken);
 		}
 
-		public override int GetItemsCount (UICollectionView collectionView, int section)
+		public override nint GetItemsCount (UICollectionView collectionView, nint section)
 		{
 			return 60;
 		}
